xhr-mixed-content.html

<!DOCTYPE html>
<html>
  <head>
    <title>Mixed-Content: XHR behavior</title>
    <meta charset='utf-8'>
    <meta name="description" content="Web-Insects">
    <link rel="author" title="Kristijan Burnik" href="burnik@chromium.org">
    <link rel="help" href="http://www.w3.org/TR/mixed-content/">
    <meta name="assert" content="Check if XHR is responding to mixed-content">
    <script src="testharness/testharness.js"></script>
    <script src="testharness/testharnessreport.js"></script>
  </head>
  <body>
    <h1>Mixed-Content: XHR events test</h1>
    <script>

      function xhrPromise(url) {
        return new Promise(function(resolve, reject) {
          var xhr = new XMLHttpRequest();

          xhr.addEventListener("error", function() {
            reject(Error("Network Error"));
          });

          xhr.addEventListener("load", function() {
            if (xhr.status != 200) {
              return reject(Error(xhr.statusText));
            }

            resolve(xhr.response);
          });

          xhr.open('GET', url, true);
          xhr.send();
        });
      }

      function runTest(url, expectation, description) {
        var xhr_test = async_test(description);

        var expectedEventTriggered = false;

        function triggerEvent(event_type) {
           console.log("Triggered event", event_type);

          expectedEventTriggered = true;
          xhr_test.step(function() {
            assert_equals(event_type, expectation,
                          "Promise should be rejected.");
          });
          xhr_test.done();
        }

        // Trigger proper event.
        xhrPromise(url).then(function() {
          triggerEvent("success");
        }, function(err) {
          triggerEvent("error");
        });


        // TODO(kristijanburnik): Once merged in this PR:
        //   https://github.com/w3c/testharness.js/pull/127
        // Pull in the new testharness code.
        // Refactor to make use of step_timeout.
        setTimeout(function(){
          xhr_test.step(function(){
            assert_true(expectedEventTriggered, "XHR event was fired.");
          })
          xhr_test.done();
        }, 300);
        /*
        xhr_test.step_timeout(function() {
          assert_true(expectedEventTriggered, "XHR event was fired.");
          xhr_test.done()
        }, 300);
        */
      }

      // Sanity check if test is run over HTTPS.
      test(function() {
        assert_equals(location.protocol, "https:",
                      "The test should be run over HTTPS.");
      }, "Check if running HTTPS", {timeout: 100});

      runTest(location.origin, "success",
              "Non mixed-content request");

      runTest(location.origin + "/nonexistingfile.html", "error",
              "Non mixed-content request with a 404 response");

      runTest("http://" + location.host, "error",
              "Mixed-content request");

    </script>
    <div id="log"></div>
  </body>
</html>
